requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
colorama==0.4.4
cycler==0.10.0
Django==3.1
django-filter==2.4.0
django-tinymce4-lite==1.8.0
iniconfig==1.1.1
joblib==0.17.0
jsmin==2.2.2
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn-pandas==2.0.3
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.2
